$_SESSION Handling Bugfix
Once you log in, some $_SESSION['user'] is populated, and previously if
any of $_SESSION['user'] was set, it wouldn't load the rest. But it
should only load the rest if it wasn't already loaded, so I added the
variable 'full_load' which indicates directly whether the full data for
the user has been populated yet or not.