Replacement of Advanced BBC with Basic Formatting
- Adds .idea to .gitignore (for PHP Storm projects)
- Renames "parser" to something slightly more intuitive and easier to
work with for an autoloader.
- Re-organizes the file in terms of when the methods are actually called
in the main source file.
- Uses a __construct to set class properties ($message, $user) rather
than passing them to a method when needed.
- Uses a get function to retrieve the final formatted message.
- Removal of all advanced BBC
- Addition of basic text formatting using * and _ for bold, and italic
text.
- Cleanup on arrays (added trailing commas, it makes additions easier).
- Added braces for if else if else conditionals to conform to coding
style, I'm up for changing this slightly though (I don't like being
super liberal with brackets)
- Changed "elseif" to "else if" (it's easier when a project uses JS and
PHP)
- Added some TODO comments of things to improve on later on.
- Changes when the parser is called. (We may actually want to modify
this so the class is instatiated in the constructor and then called for
each message - ie: setMessage($message) and then it already uses
getMessage(); may be more intuitive in the long run).

fixes 15